pyproject.toml
[project]
name = "rag-implementation-gemini"
version = "0.1.0"
description = "A simple RAG (Retrieval-Augmented Generation) demo using Gemini API and uv"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.3.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.1",
    "PyPDF2>=3.0.0",
    "streamlit>=1.28.0"
]
authors = [
    {name = "Sumanta Swain", email = "sumanta_swain@epam.com"}
]

[dependency-groups]
dev = ["pytest"]

[project.scripts]
rag-gemini = "main:main"
